Use built-in widgets to layout Landowner Review
Remove references to the clumsy use of
`lib/components/relative_sized_box.dart` and
`lib/util/relative_size.dart` in favor of built-in widgets. Do this to
simplify code and also take advantage of baked-in responsiveness.

However this causes landscape mode issues where the listed areas are
in a barely visible sliver due to overly large contact info text.