Instantiate singleton `SharedPreferences`
Since Shared Preferences is a singleton keep a single reference to it in
order to allow it to manage all the async code instead of littering the
codebase with `async` and `await`. Use an inherited widget to wrap the
entire `MaterialApp` in order to allow it to be accessed from anywhere
in the tree since in the futures font size setting will likely be need
to be available from each screen.