Add checklist DOCX template
Note that the template must have content control tags with `text` as
tag. To access content control tags in MS Word 2013, go to the Developer
ribbon, click `Design Mode`, and place a `Plain Text Content
Control`. Then while selecting the newly placed content control, select
the `Properties` option in the `Developer` ribbon. The two important
fields here are `title` (which is how the Dart library `docx_template`
will locate the content control) and `tag` (which must be set to `text`
for `docx_template` to be able to recognize the tag).