FIX: Read in checklist DOCX template a single time
Instead of creating a separate `DOCXConverter` for each separate
`AreaReview` screen, make a single provided `DOCXConverter` to avoid
having to reread the checklist docx file each time the user clicks the
`AreaReview` screen for a different area.

Consequence: This means that `AreaReview` can be made stateless again
sometime in the future.